fixed: every id type changed to int64

+++ domain/comments.go
@@ -2,9 +2,9 @@ package domain
 
 type CommentUsecase interface {
 	CreateComment(comm *Comment) error
-	GetCommentsByPostID(id int) ([]*Comment, error)
+	GetCommentsByPostID(id int64) ([]*Comment, error)
 }
 type CommentRepo interface {
 	CreateComment(comm *Comment) error
-	GetCommentsByPostID(id int) ([]*Comment, error)
+	GetCommentsByPostID(id int64) ([]*Comment, error)
 }
